Atomistic_modeling
These files are on atomistic modelling of silicon and silica using Lammps. 
Some auxilliary python scripts to assist in the work of material modeling are also included. The python script to read cif files can convert a cif geometry file of a substance into a data file with coordinates of atoms understandable by lammps. the other python scripts help to create the lammps input data. 
There are other matlab(*.m) and python scripts which help in visualizing and processing corresponding lammps output data files. 
Lammps scripts created restart files that are also included and can be used to restart the system from previous run.
There has to be a dump folder for running the lammps scripts to store the output images and atom coordinates from the run. 
Some results are also included. The latest project is in progress and some initial results have been committed here. It will be updated soon.