Composites project code
To be run in matlab/octave. 
Analyses response and calculates properties of a composite from its constituent data (eg- elastic properties of fibre and matrix). Each layer in the composite should be made of only two constituents (a fibre and a composite). These may be different for different layers.
Use consistent units. Adviced to use GPa and mm for stress and length units.